+++ klause/src/commonTest/kotlin/com/eignex/klause/lp/cut/FlowCoverSeparatorTest.kt
@@ -135,7 +135,7 @@ class FlowCoverSeparatorTest {
             fun feasible(): Boolean {
                 for (f in p.factors.filterIsInstance<Linear>()) {
                     var s = 0L
-                    for (i in f.vars.indices) s += f.coeffs[i].toLong() * point[f.vars[i]]
+                    for (i in f.vars.indices) s += f.coeffs[i] * point[f.vars[i]]
                     if (s > f.bound) return false // every factor is LE here
                 }
                 return true
